requirements.txt
albumentations==1.3.1
einops==0.7.0
ImageHash==4.3.1
matplotlib==3.8.2
MedSimilarity==1.0.0
nibabel==5.2.0
numpy==1.26.3
opencv_python_headless==4.9.0.80
Pillow==10.2.0
pydantic==2.5.3
sentence_transformers==2.2.2
setuptools==59.6.0
torch==2.1.2
torchmetrics==1.3.0
torchvision==0.16.2
